# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toilet Finder API",
        "description": "Office Toilet Finder API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "employee_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "employee_id": {
                                        "type": "string",
                                        "example": "EMP123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "employee_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/maintenance/report": {
            "post": {
                "tags": [
                    "Maintenance"
                ],
                "summary": "Submit a maintenance report",
                "operationId": "793eb83d386f0d559d27caf3219fe470",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "toilet_id",
                                    "issue_type",
                                    "description",
                                    "priority"
                                ],
                                "properties": {
                                    "toilet_id": {
                                        "type": "integer"
                                    },
                                    "issue_type": {
                                        "type": "string",
                                        "enum": [
                                            "plumbing",
                                            "cleaning",
                                            "repair",
                                            "other"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "urgent"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "report": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "toilet_id": {
                                                    "type": "integer"
                                                },
                                                "issue_type": {
                                                    "type": "string"
                                                },
                                                "priority": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/washrooms": {
            "get": {
                "tags": [
                    "Washrooms"
                ],
                "summary": "Get all washrooms with their toilets",
                "operationId": "7c7a224ce5da6b0380355e6118c84bb5",
                "responses": {
                    "200": {
                        "description": "List of washrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "floor": {
                                                "type": "string"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "male",
                                                    "female",
                                                    "unisex"
                                                ]
                                            },
                                            "is_operational": {
                                                "type": "boolean"
                                            },
                                            "available_toilets": {
                                                "type": "integer"
                                            },
                                            "toilets": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "number": {
                                                            "type": "string"
                                                        },
                                                        "is_occupied": {
                                                            "type": "boolean"
                                                        },
                                                        "occupied_by": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "time_remaining": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API Endpoints for user authentication"
        },
        {
            "name": "Maintenance",
            "description": "Maintenance"
        },
        {
            "name": "Washrooms",
            "description": "Washrooms"
        }
    ]
}